Move palette into tailwind; extract projects list
