+++ src/components/Card.tsx
@@ -1,5 +1,4 @@
 import type { FC, PropsWithChildren } from "react";
-import palette from "../palette";
 
 interface CardProps {
   className?: string;
@@ -8,12 +7,7 @@ const Card: FC<CardProps> = ({
   children,
   className,
 }: PropsWithChildren<CardProps>) => (
-  <div
-    className={`box-border p-5 ${className ?? ""}`}
-    style={{ backgroundColor: palette.card }}
-  >
-    {children}
-  </div>
+  <div className={`bg-card box-border p-5 ${className ?? ""}`}>{children}</div>
 );
 
 export default Card;
