+++ src/components/Quiet.tsx
@@ -1,9 +1,8 @@
 import type { FC, PropsWithChildren } from "react";
-import palette from "../palette";
 
 export interface QuietProps {}
 const Quiet: FC<QuietProps> = ({ children }: PropsWithChildren<QuietProps>) => (
-  <i style={{ color: palette.subtle }}>{children}</i>
+  <i className="text-subtle">{children}</i>
 );
 
 export default Quiet;
