+++ src/layouts/Footer.astro
@@ -2,14 +2,13 @@
 import ExternalLink from "../components/ExternalLink";
 import Github from "../components/icons/Github";
 import Three from "./Three.astro";
-import palette from "../palette";
 ---
 
-<nav class="sticky top-0" style={{ backgroundColor: palette.footer }}>
+<nav class="sticky top-0 bg-footer">
   <Three>
     <div class="flex justify-end gap-5">
       <ExternalLink href="https://www.github.com/jbarlo/">
-        <span class=`text-[${palette.text}] hover:text-white`>
+        <span class="text-text hover:text-white">
           <Github />
         </span>
       </ExternalLink>
