+++ src/pages/index.astro
@@ -3,7 +3,6 @@ import Layout from "../layouts/Layout.astro";
 import Header from "../layouts/Header.astro";
 import Three from "../layouts/Three.astro";
 import Footer from "../layouts/Footer.astro";
-import palette from "../palette";
 
 // TODO first 3
 const posts = [
@@ -16,14 +15,10 @@ const posts = [
 ];
 ---
 
-<Layout title="home | barlo" overflowColor={palette.footer}>
-  <div style={{ backgroundColor: palette.bg, color: palette.text }}>
+<Layout title="home | barlo">
+  <div class="bg-bg text-text">
     <div class="flex flex-col min-h-screen leading-relaxed">
-      <Header
-        backgroundColor={palette.bg}
-        titleColor={palette.title}
-        subdomainColor={palette.subdomain}
-      />
+      <Header />
       <Three>
         <div class="flex flex-col gap-2">
           <h2 class="text-2xl">recent</h2>
