+++ src/projects.ts
@@ -1,0 +1,67 @@
+const projects: {
+  id: string;
+  label: string;
+  lastUpdated: Date;
+  imgUrl: string;
+  imgAlt: string;
+  github?: string;
+}[] = [
+  {
+    id: "test1",
+    label: "Test 1",
+    lastUpdated: new Date(),
+    imgUrl:
+      "https://upload.wikimedia.org/wikipedia/commons/9/99/Brooks_Chase_Ranger_of_Jolly_Dogs_Jack_Russell.jpg",
+    imgAlt: "is dog",
+  },
+  {
+    id: "test2",
+    label: "Test 2",
+    lastUpdated: new Date(),
+    imgUrl:
+      "https://upload.wikimedia.org/wikipedia/commons/9/99/Brooks_Chase_Ranger_of_Jolly_Dogs_Jack_Russell.jpg",
+    imgAlt: "is dog",
+  },
+  {
+    id: "test3",
+    label: "Test 3",
+    lastUpdated: new Date(),
+    imgUrl:
+      "https://upload.wikimedia.org/wikipedia/commons/9/99/Brooks_Chase_Ranger_of_Jolly_Dogs_Jack_Russell.jpg",
+    imgAlt: "is dog",
+  },
+  {
+    id: "test1",
+    label: "Test 1",
+    lastUpdated: new Date(),
+    imgUrl:
+      "https://upload.wikimedia.org/wikipedia/commons/9/99/Brooks_Chase_Ranger_of_Jolly_Dogs_Jack_Russell.jpg",
+    imgAlt: "is dog",
+  },
+  {
+    id: "test4",
+    label: "Test 4",
+    lastUpdated: new Date(),
+    imgUrl:
+      "https://upload.wikimedia.org/wikipedia/commons/9/99/Brooks_Chase_Ranger_of_Jolly_Dogs_Jack_Russell.jpg",
+    imgAlt: "is dog",
+  },
+  {
+    id: "test3",
+    label: "Test 3",
+    lastUpdated: new Date(),
+    imgUrl:
+      "https://upload.wikimedia.org/wikipedia/commons/9/99/Brooks_Chase_Ranger_of_Jolly_Dogs_Jack_Russell.jpg",
+    imgAlt: "is dog",
+  },
+  {
+    id: "test1",
+    label: "Test 1",
+    lastUpdated: new Date(),
+    imgUrl:
+      "https://upload.wikimedia.org/wikipedia/commons/9/99/Brooks_Chase_Ranger_of_Jolly_Dogs_Jack_Russell.jpg",
+    imgAlt: "is dog",
+  },
+];
+
+export default projects;
